Restrict high-kingdom to holders of 2 de-jure kingdoms
The option is visible when a vassal holds more than 1 de-jure Kingdom
in the Empire of their primary title.

But to actually do it you need to hold all of the land in the de-jure
Empire.

The new condition lets you know that it's an option you can work towards,
and stops it appearing after you've already made the High Kingdom,
as there's only one de-jure Kingdom in the de-jure Empire.