Convert High Kingdom creation to an Event
This is the closest you can get to a function call, which I'm using to
reduce code duplication.

This isn't a massive reduction in code because of the event definition
overhead, but it's easier to make changes to how High Kingdoms happen now,
as there's only two places that need to be kept in sync.